학생 db entity 추가

+++ database/migrations/2017_05_06_102414_create_course_student_table.php
@@ -15,7 +15,7 @@ class CreateCourseStudentTable extends Migration
     {
         Schema::create('course_student', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('fee')->unsigned();
+            $table->double('fee',8,2)->unsigned();
             $table->integer('student_id')->unsigned();
             $table->integer('course_id')->unsigned();
